requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==46.0.1
defusedxml==0.7.1
diff-match-patch==20241021
dj-rest-auth==7.0.1
Django==5.2.4
django-allauth==65.11.2
django-cors-headers==4.7.0
django-filter==25.1
django-google-sso==7.1.0
django-import-export==4.3.9
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
drf-yasg==1.21.10
email-validator==2.3.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
gunicorn==23.0.0
haversine==2.9.0
httplib2==0.31.0
idna==3.10
inflection==0.5.1
loguru==0.7.3
mailersend==0.5.6
msgpack==1.1.1
oauthlib==3.3.1
packaging==25.0
paho-mqtt==2.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==68.2.2
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0
win32_setctime==1.2.0
